// frontend/src/api/model/addressDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { LocationDto } from './locationDto';

/**
 * объект
 */
export interface AddressDto {
  id: number;
  street?: string;
  zipCode?: string;
  town?: LocationDto;
}